Renamed some shared examples files
