+++ spec/api/v1/answers_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
-require 'shared_examples/controllers/api/v1/api_authenticated'
+require 'shared_examples/controllers/api/v1/api_shared_examples'
 
 describe Api::V1::AnswersController, type: :controller do
   describe 'GET #index' do
