+++ spec/models/comment_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
-require 'shared_examples/models/broadcastable_spec'
+require 'shared_examples/models/broadcastable'
 
 RSpec.describe Comment, type: :model do
   it_should_behave_like 'broadcastable'
