+++ spec/shared_examples/controllers/api/v1/api_shared_examples.rb
