+++ spec/shared_examples/controllers/voted.rb
