+++ spec/shared_examples/models/attachable.rb
