+++ spec/shared_examples/models/broadcastable.rb
