+++ spec/shared_examples/models/commentable.rb
