+++ spec/shared_examples/models/votable.rb
